+++ src/devices/Vcnl4040/tests/Internal/AlsLowInterruptThresholdRegisterTest.cs
@@ -31,16 +31,16 @@ namespace Iot.Device.Vcnl4040.Tests.Internal
         [Theory]
         [InlineData(0b0101_0101_1010_1010, 0b0101_0101, 0b1010_1010)]
         [InlineData(0b1010_1010_0000_0101, 0b1010_1010, 0b0000_0101)]
-        public void Write(int threshold, byte expectedThresholdHighByte, byte expectedThresholdLowByte)
+        public void Write(ushort threshold, byte expectedThresholdHighByte, byte expectedThresholdLowByte)
         {
-            PropertyWriteTest<AlsLowInterruptThresholdRegister, int>(initialRegisterLowByte: 0x00,
-                                                                     initialRegisterHighByte: 0x00,
-                                                                     testValue: threshold,
-                                                                     expectedLowByte: expectedThresholdLowByte,
-                                                                     expectedHighByte: expectedThresholdHighByte,
-                                                                     commandCode: (byte)CommandCode.ALS_THDL,
-                                                                     registerPropertyName: nameof(AlsLowInterruptThresholdRegister.Level),
-                                                                     registerReadsBeforeWriting: false);
+            PropertyWriteTest<AlsLowInterruptThresholdRegister, ushort>(initialRegisterLowByte: 0x00,
+                                                                        initialRegisterHighByte: 0x00,
+                                                                        testValue: threshold,
+                                                                        expectedLowByte: expectedThresholdLowByte,
+                                                                        expectedHighByte: expectedThresholdHighByte,
+                                                                        commandCode: (byte)CommandCode.ALS_THDL,
+                                                                        registerPropertyName: nameof(AlsLowInterruptThresholdRegister.Level),
+                                                                        registerReadsBeforeWriting: false);
         }
     }
 }
